refactor: masterTableViewController DataSource 분리 #2
- 상속하지 않는 클래스의 경우 final 키워드 추가
